fix: consume main view model dialog everywhere
